+++ tests/testthat/test-pal.R
@@ -1,9 +1,9 @@
 context("palettes")
 
-test_that("trekpal returns as expected", {
-  expect_error(trekpal("a"), "`id` is not a predefined palette name.")
-  x <- trekpal()
+test_that("trek_pal returns as expected", {
+  expect_error(trek_pal("a"), "Invalid palette name.")
+  x <- trek_pal()
   expect_is(x, "character")
-  expect_equal(length(x), 53)
-  expect_equal(length(trekpal("lcarscolors")), 31)
+  expect_equal(length(x), 39)
+  expect_equal(length(trek_pal("lcars_series")), 31)
 })