+++ NEWS.md
@@ -1,4 +1,6 @@
 # trekcolors 0.1.0
 
 * Added package scaffolding.
-* Added initial color palettes.
+* Added initial color palettes
+* Added related functions.
+* Added documentation and tests.